Hotel Reservation system(web application)
This is the latest version of the file. This is a hotel reservation system called "Ceylon Haven". It is mainly used by clients. It has a navigation bar that allows you to switch to several pages. It can be used by both clients/administrators. Currently, only the frontend is available. HTML, CSS, and ReactJS are used.